fix!: prevent racing of event stream connection between devices (#34)
This adds the capability to provide the name of the webex device this crate uses.
If existing, the latest device with that name will be used. If it doesn't, a new
one will be created with that name.

A default name is provided to ensure backwards compatibility while still fixing
the race condition with official clients.

BREAKING CHANGE: existing webex devices that were not created by this crate
will no longer be selected when creating the event stram. This is the intent
of this fix but represents a change in behaviour.